+++ projects/swimlane/ngx-graph/src/lib/models/node.model.ts
@@ -16,8 +16,15 @@ export interface Node {
   label?: string;
   data?: any;
   meta?: any;
+  layoutOptions?: any;
+  parentId?: string;
+  hidden?: boolean;
 }
 
 export interface ClusterNode extends Node {
+  childNodeIds?: string[];
+}
+
+export interface CompoundNode extends Node {
   childNodeIds?: string[];
 }