+++ CHANGELOG.md
@@ -1,9 +1,23 @@
 # Changelog
 
+## 8.2.1
+
+- Bug: When using `deferDisplayUntilPosition` compound nodes and clusters may not be displayed (#511)
+
+## 8.2.0
+
+- Feature: Update Angular 16 as a peer dependency (#499)
+- Feature: Use `deferDisplayUntilPosition` Input to display nodes after position returned by layout (#509)
+
+## 8.1.0
+
+- Feature: Support Elk Compound Nodes (#502) (#506)
+- Bug: Fix issues with the build (#507)
+
 ## 8.0.3
 
 - Chore: Bump d3 dependencies (#477)
-- Fix an issue where fixing a node position would not work (#463)
+- Bug: Fix an issue where fixing a node position would not work (#463)
 
 ## 8.0.2